change board_detection to board_vision
